refactor: 폴더 구조 변경 및 사용하지 않는 파일 삭제
