+++ README.md
@@ -1,3 +1,7 @@
+# javascript-lunch
+
+우아한테크코스 레벨2 점심 뭐 먹지 미션
+
 # Getting Started with Create React App
 
 This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
