--- README.old.md
@@ -1,2 +1,0 @@
-# javascript-lunch
-우아한테크코스 레벨2 점심 뭐 먹지 미션
