+++ src/components/RestaurantDetail.tsx
@@ -2,7 +2,7 @@ import { Restaurant } from '../types/restaurant';
 import React from 'react';
 import styled from 'styled-components';
 import { CATEGORY_IMAGE_MAP } from '../constants';
-import { textBody, textSubtitle } from '../mixin';
+import { textBody, textSubtitle } from '../style/mixin';
 
 interface Props {
   closeModal: () => void;
